server/requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Werkzeug==2.3.8
openai-whisper==20231117
google-cloud-translate==3.11.0
gTTS==2.3.2
torch==2.1.0
websockets==11.0.3
